ai-smart-assistant/backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
sqlalchemy>=2.0.23
pydantic>=2.8.2
python-dotenv>=1.0.0
openai>=1.40.0
aiohttp>=3.10.0
pinecone-client>=4.1.2